Move branch related data into their own module
